requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
opencv-python-headless==4.9.0.80
